Rename parserfilegenerator.rb to parser_file_generator.rb
