+++ lib/racc/static.rb
@@ -1,5 +1,5 @@
 require 'racc'
 require 'racc/parser'
 require 'racc/grammar_file_parser'
-require 'racc/parserfilegenerator'
+require 'racc/parser_file_generator'
 require 'racc/log_file_generator'
